move everything ProfileView.vue, which was previously a temporary view to check the ChallengeInteraction component
