+++ client/src/router/index.ts
@@ -8,20 +8,7 @@ import FriendView from '@/views/FriendView.vue'
 import BlingoView from '@/views/BlingoView.vue'
 import PreferenceView from '@/views/PreferenceView.vue'
 import ProfileView from '@/views/ProfileView.vue'
-import UserProfileView from '@/views/UserProfileView.vue'
-import type { RouteLocationNormalized } from 'vue-router'
 
-export const routes = [
-  // ... 其他路由
-  {
-    path: '/profile/:username?',
-    name: 'user-profile',
-    component: UserProfileView,
-    props: (route: RouteLocationNormalized) => ({
-      username: route.params.username as string | undefined,
-    }),
-  },
-]
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
